fix: :sparkles: make new algorithm logic the main logic, remove old code
