+++ __test__/integration/integrity.test.js
@@ -11,7 +11,7 @@ describe("Testing TRAPIQueryHandler Module", () => {
         const queryHandler = new TRAPIQueryHandler.TRAPIQueryHandler({}, undefined, undefined, true);
         const query = JSON.parse(fs.readFileSync(path.join(example_foler, 'increased_urinary_glycerol_affects_glycerol.json')));
         queryHandler.setQueryGraph(query.message.query_graph);
-        await queryHandler.query_2();
+        await queryHandler.query();
         const res = queryHandler.getResponse();
         expect(res.message.knowledge_graph.nodes).toHaveProperty('CHEBI:17754');
     })
@@ -20,7 +20,7 @@ describe("Testing TRAPIQueryHandler Module", () => {
         const queryHandler = new TRAPIQueryHandler.TRAPIQueryHandler({}, undefined, undefined, true);
         const query = JSON.parse(fs.readFileSync(path.join(example_foler, 'FDFM_caused_by_ACDY5.json')));
         queryHandler.setQueryGraph(query.message.query_graph);
-        await queryHandler.query_2();
+        await queryHandler.query();
         const res = queryHandler.getResponse();
         expect(res.message.knowledge_graph.nodes).toHaveProperty('NCBIGene:111');
     })
@@ -30,7 +30,7 @@ describe("Testing TRAPIQueryHandler Module", () => {
         const queryHandler = new TRAPIQueryHandler.TRAPIQueryHandler({}, undefined, undefined, true);
         const query = JSON.parse(fs.readFileSync(path.join(example_foler, 'chemicals_targeting_IL1_Signaling_Pathway.json')));
         queryHandler.setQueryGraph(query.message.query_graph);
-        await queryHandler.query_2();
+        await queryHandler.query();
         const res = queryHandler.getResponse();
         expect(res.message.knowledge_graph.nodes).toHaveProperty('CHEBI:3962');
     })
