+++ src/index.js
@@ -3,7 +3,7 @@ var path = require('path');
 const BatchEdgeQueryHandler = require('./batch_edge_query');
 const QueryGraph = require('./query_graph');
 const KnowledgeGraph = require('./graph/knowledge_graph');
-const QueryResults = require('./query_results_2');
+const QueryResults = require('./query_results');
 const InvalidQueryGraphError = require('./exceptions/invalid_query_graph_error');
 const debug = require('debug')('bte:biothings-explorer-trapi:main');
 const Graph = require('./graph/graph');
@@ -74,26 +74,7 @@ exports.TRAPIQueryHandler = class TRAPIQueryHandler {
    * @private
    * @param {object} queryGraph - TRAPI Query Graph Object
    */
-  _processQueryGraph(queryGraph) {
-    try {
-      let queryGraphHandler = new QueryGraph(queryGraph);
-      let res = queryGraphHandler.createQueryPaths();
-      this.logs = [...this.logs, ...queryGraphHandler.logs];
-      return res;
-    } catch (err) {
-      if (err instanceof InvalidQueryGraphError) {
-        throw err;
-      } else {
-        throw new InvalidQueryGraphError();
-      }
-    }
-  }
-
-  /**
-   * @private
-   * @param {object} queryGraph - TRAPI Query Graph Object
-   */
-   async _processQueryGraph_2(queryGraph) {
+   async _processQueryGraph(queryGraph) {
     try {
       let queryGraphHandler = new QueryGraph(queryGraph);
       let res = await queryGraphHandler.calculateEdges();
@@ -119,41 +100,18 @@ exports.TRAPIQueryHandler = class TRAPIQueryHandler {
     return handlers;
   }
 
-  async query() {
-    this._initializeResponse();
-    debug('start to load metakg.');
-    const kg = this._loadMetaKG(this.smartapiID, this.team);
-    debug('metakg successfully loaded');
-    let queryPaths = this._processQueryGraph(this.queryGraph);
-    debug(`query paths constructed: ${JSON.stringify(queryPaths)}`);
-    const handlers = this._createBatchEdgeQueryHandlers(queryPaths, kg);
-    debug(`Query depth is ${Object.keys(handlers).length}`);
-    for (let i = 0; i < Object.keys(handlers).length; i++) {
-      debug(`Start to query depth ${i + 1}`);
-      let res = await handlers[i].query(handlers[i].qEdges);
-      debug(`Query for depth ${i + 1} completes.`);
-      this.logs = [...this.logs, ...handlers[i].logs];
-      if (res.length === 0) {
-        return;
-      }
-      debug('Start to notify subscribers now.');
-      handlers[i].notify(res);
-      debug(`Updated TRAPI knowledge graph using query results for depth ${i + 1}`);
-    }
-  }
-
   _createBatchEdgeQueryHandlersForCurrent(currentEdge, kg) {
     let handler = new BatchEdgeQueryHandler(kg, this.resolveOutputIDs);
     handler.setEdges(currentEdge);
     return handler;
   }
 
-  async query_2() {
+  async query() {
     this._initializeResponse();
     debug('Start to load metakg.');
     const kg = this._loadMetaKG(this.smartapiID, this.team);
     debug('MetaKG successfully loaded!');
-    let queryEdges = await this._processQueryGraph_2(this.queryGraph);
+    let queryEdges = await this._processQueryGraph(this.queryGraph);
     debug(`(3) All edges created ${JSON.stringify(queryEdges)}`);
     const manager = new EdgeManager(queryEdges);
     while (manager.getEdgesNotExecuted()) {
@@ -163,7 +121,7 @@ exports.TRAPIQueryHandler = class TRAPIQueryHandler {
       let handler = this._createBatchEdgeQueryHandlersForCurrent(current_edge, kg);
       debug(`(5) Executing current edge >> "${current_edge.getID()}"`);
       //execute current edge query
-      let res = await handler.query_2(handler.qEdges);
+      let res = await handler.query(handler.qEdges);
       this.logs = [...this.logs, ...handler.logs];
       if (res.length === 0) {
         debug(`(X) Terminating..."${current_edge.getID()}" got 0 results.`);
